requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
Django==2.1.5
djangorestframework==3.9.0
kombu==4.2.2.post1
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
vine==1.2.0
